Posts page shows posts
Restructured a little bit, added some ui components.
