+++ models/post.js
@@ -15,6 +15,7 @@ module.exports = (sequelize, DataTypes) => {
 
   Post.associate = (models) => {
     Post.belongsTo(models.user);
+    Post.belongsTo(models.weed);
   };
 
   return Post;
