+++ react-client/src/components/MainPage.js
