+++ react-client/src/components/Nav.js
@@ -12,6 +12,7 @@ const useStyles = makeStyles(theme => ({
   },
   title: {
     flexGrow: 1,
+    color: "white"
   },
 }));
 
@@ -26,7 +27,7 @@ const Nav = (props) => {
           <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="Menu">
             <MenuIcon />
           </IconButton>
-          <Typography variant="h6" className={classes.title}>
+          <Typography variant="h6" className={classes.title} >
             TokeTalk
           </Typography>
           <Button className={classes.menuButton} color="inherit" aria-label="Menu" onClick={props.home}>Home</Button>
