+++ react-client/src/components/NewCharacter.css
