+++ react-client/src/components/NewCharacter.js
