+++ react-client/src/components/WeedFooter.js
