+++ react-client/src/components/ui/Box.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Box = (props) => {
+  return (
+    <div {...props} >
+      {props.children}
+    </div>
+  )
+}
+
+export default Box;
